feat: добавит компонент мемоизированный компонент header
